Add a test. Move document stories under documents.
