+++ src/lib/components/common/tests/Button.test.ts
@@ -1,6 +1,5 @@
-import { describe, it, expect, vi } from "vitest";
+import { describe, it, expect } from "vitest";
 import { render, screen } from "@testing-library/svelte";
-import { userEvent } from "@testing-library/user-event";
 
 import Button from "../Button.svelte";
